requirements.txt
asgiref==3.9.1
Django==4.2
django-filter==25.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
